refactor: move models to domain-specific folders
