+++ src/domains/product/types.ts
