+++ src/domains/user/constants.ts
